Restructuring of the Id/Key/Owner implementation used for all collection keys. This replaces the use of the term Id in most cases with Key. Id is the lower bits of Key only. Owner is now always a value from 1 to 255 (zero if unknown/unset).

+++ trunk/Common/DataStructures/PlayerSettings.cs
@@ -36,7 +36,7 @@ namespace Nova.Common
     {
         public string RaceName; // The path & file name of the race.
         public string AiProgram; // The path & file name of the AI application or "Human"
-        public int PlayerNumber; // The order number of the player from 1 - Global.MaxPlayers        
+        public ushort PlayerNumber; // The order number of the player from 1 - Global.MaxPlayers        
     
         /// <summary>
         /// Default constructor. 
@@ -65,7 +65,7 @@ namespace Nova.Common
                             AiProgram = subnode.FirstChild.Value;
                             break;
                         case "playernumber":
-                            PlayerNumber = int.Parse(subnode.FirstChild.Value, System.Globalization.CultureInfo.InvariantCulture);
+                            PlayerNumber = ushort.Parse(subnode.FirstChild.Value, System.Globalization.CultureInfo.InvariantCulture);
                             break;
                     }
                 }
